Cherry-pick PRs and release notes for v2.4.3 (#1016)
* double-check unpackedBuffer created in DWA uncompress


* compute Huf codelengths using 64 bit to prevent shift overflow


* Avoid overflow in calculateNumTiles when size=MAX_INT (#825)

* Avoid overflow in calculateNumTiles when size=MAX_INT


* Compute level size with 64 bits to avoid overflow


* More efficient handling of filled channels reading tiles with scanline API (#830)

* refactor channel filling in InputFile API with tiled source


* handle edge-case of empty framebuffer


* fix undefined behavior: ignore unused bits in B44 mode detection (#832)



* Fix overflow computing deeptile sample table size (#861)



* sanity check ScanlineInput bytesPerLine instead of lineOffset size (#863)



* Release notes for v2.4.3


* Bump version for v2.4.3


* reduce size limit for scanline files; prevent large chunkoffset allocations (#824)

* reduce size limit for scanline files; protect against large chunkoffset allocations


* bugfix for memory limit changes


* rearrange chunkoffset test to protect bytesperline table too


* remove extraneous function declaration; tidy comments


* Change v2.4.3 release date to May 17, and clean up urls

+++ OpenEXR/IlmImf/ImfDeepTiledInputFile.cpp
@@ -976,8 +976,8 @@ DeepTiledInputFile::initialize ()
     for (size_t i = 0; i < _data->tileBuffers.size(); i++)
         _data->tileBuffers[i] = new TileBuffer ();
 
-    _data->maxSampleCountTableSize = _data->tileDesc.ySize *
-                                     _data->tileDesc.xSize *
+    _data->maxSampleCountTableSize = static_cast<size_t>(_data->tileDesc.ySize) *
+                                     static_cast<size_t>(_data->tileDesc.xSize) *
                                      sizeof(int);
 
     _data->sampleCountTableBuffer.resizeErase(_data->maxSampleCountTableSize);
